Adding delete button to tasks; fixes #45.

+++ inthe_am/taskmanager/static/modules/controllers/task.js
@@ -2,13 +2,16 @@ var controller = Ember.ObjectController.extend({
   needs: ['tasks'],
   actions: {
     'complete': function(){
-      var self = this;
-      this.get('model').destroyRecord().then(function(){
-        self.get('controllers.tasks').refresh();
-        self.transitionToRoute('tasks');
-      }, function(){
-        alert("An error was encountered while marking this task completed.");
-      });
+      var result = confirm("Are you sure you would like to mark this task as completed?");
+      if(result) {
+        var self = this;
+        this.get('model').destroyRecord().then(function(){
+          self.get('controllers.tasks').refresh();
+          self.transitionToRoute('tasks');
+        }, function(){
+          alert("An error was encountered while marking this task completed.");
+        });
+      }
     },
     'delete_annotation': function(description) {
       var model = this.get('model');
@@ -49,21 +52,20 @@ var controller = Ember.ObjectController.extend({
       });
     },
     'delete': function(){
-      var url = this.store.adapterFor('task').buildURL('task', this.get('uuid')) + 'delete/';
-      $.ajax({
-        url: url,
-        dataType: 'json',
-        statusCode: {
-          200: function(){
+      var result = confirm("Are you sure you would like to delete this task?");
+      if(result) {
+        var self = this;
+        var url = this.store.adapterFor('task').buildURL('task', this.get('uuid')) + 'delete/';
+        $.ajax({
+          url: url,
+          dataType: 'json',
+          success: function(){
             self.get('model').unloadRecord();
             self.get('controllers.tasks').refresh();
             self.transitionToRoute('tasks');
-          },
-          501: function(){
-            alert("Deleting tasks is currently unimplemented");
           }
-        }
-      });
+        });
+      }
     }
   }
 });
